fix: consolidating the modals, removing the need for a second modal for deposit
